Fixes for errors and warnings building for Linux.

+++ src/device.c
@@ -447,7 +447,7 @@ static int handler_signal(const char *path, const char *types, lo_arg **argv,
             if (id_map_index < 0) {
 #ifdef DEBUG
                 printf("no local instances available for global instance id"
-                       " %llu\n", global_id);
+                       " %llu\n", (unsigned long long)global_id);
 #endif
                 return 0;
             }
@@ -561,7 +561,7 @@ static int handler_signal(const char *path, const char *types, lo_arg **argv,
                     if (id_map_index < 0) {
 #ifdef DEBUG
                         printf("no local instances available for global"
-                               " instance id %llu\n", global_id);
+                               " instance id %llu\n", (unsigned long long)global_id);
 #endif
                         return 0;
                     }
