+++ src/signal.c
@@ -538,7 +538,7 @@ int mapper_signal_instance_with_global_id(mapper_signal sig, mapper_id global_id
             /* TODO: Once signal groups are explicit, allow re-mapping to
              * another instance if possible. */
             trace("Signal %s has no instance %llu available.\n", sig->name,
-                  map->local);
+                  (unsigned long long)map->local);
             return -1;
         }
         else if (!si->is_active) {
@@ -599,7 +599,7 @@ int mapper_signal_instance_with_global_id(mapper_signal sig, mapper_id global_id
         si = find_instance_by_id(sig, map->local);
         if (!si) {
             trace("Signal %s has no instance %llu available.", sig->name,
-                  map->local);
+                  (unsigned long long)map->local);
             return -1;
         }
         if (si) {
